Move some keywords
