+++ as3lib/_toplevel/__init__.py
@@ -1813,3 +1813,45 @@ def parseInt(str: String = undefined, radix: uint = 0):
 def unescape(str):
    raise NotImplementedError
 
+
+# Keyword functions
+def delete(obj):
+   # TODO: Other functionality of delete
+   del obj
+   return true
+
+
+def each(iterable):
+   '''
+   Replacement for 'for each ...'
+
+   Used like 'for i in each(<variable>):'
+   '''
+   if hasattr(iterable, '__each__'):
+      return iterable.__each__()
+   if hasattr(iterable, 'values') and callable(iterable.values):
+      return iterable.values()
+   # Do it a bit jank because python and actionscript differ in the way that
+   # iterating is done.
+   return (iterable[i] for i in range(len(iterable)))
+
+
+# Operations
+def stricteq(obj1, obj2):
+   if isinstance(obj1, Number) and obj1._is_nan() and isinstance(obj2, Number) and obj2._is_nan():
+      return true
+   return Boolean(type(obj1) == type(obj2) and obj1 == obj2)
+
+
+def strictne(obj1, obj2):
+   return Boolean(type(obj1) != type(obj2) or obj1 != obj2)
+
+
+# Helpers
+def as3_enumerate(iterable):
+   '''
+   Python enumerate function for AS3 objects. AS3 objects use a custom
+   implementation of __iter__ which breaks the builtin enumerate function.
+   '''
+   return ((i, iterable[i]) for i in iterable)
+
